Removed redundant web config in RequisitionApplication.java.
Took out SecretKeyProvider as a stand-alone class instead of a nested class in JwtService to avoid circular dependency

Added @Component annotation to LoginService to make it recognizable by Autowire

App currently runs but returns a "GET method not found" error message. Next step is to modify db structure with added password, create a stored procedure and figure out why GET is not working.